FA report: The floating menu that appears next to rows now uses the CSS property 'visibility' to toggle between visible and hidden
This prevents an issue with the previous way to show and hide the menu (switching between values of CSS property 'display') which causes an almost empty cell to appear and disappear at the end of the currently-moused-over row. This causes the entire table to shift slightly when the menu appears.

This commit uses an alternative implementation to show and hide the menu, though it does not completely solved the above-mentioned annoyance. This implementation works around the issue by always displaying the almost-empty cell, so the entire table does not appear to shift slightly when the menu appears and disappears.